+++ include/AST.h
@@ -5,8 +5,46 @@
 
 namespace panther{
 	namespace AST{
-		
-		struct NodeID{ uint32_t id; };
+
+		enum class Kind{
+			VarDecl,
+			Type,
+
+			// tokens
+			Ident,
+			Literal,
+		};
+
+
+		struct Node{
+			struct ID{ // typesafe identifier
+				uint32_t id;
+				explicit ID(uint32_t _id) noexcept : id(_id) {};
+			};
+
+			Kind kind;
+
+			union {
+				uint32_t index;
+				Token::ID token;
+			};
+
+			Node(Kind node_kind, uint32_t node_index) : kind(node_kind), index(node_index){};
+			Node(Kind node_kind, Token::ID token_id) : kind(node_kind), token(token_id){};
+		};
+
+
+
+		struct VarDecl{
+			Node::ID ident;
+			Node::ID type;
+			Node::ID expr;
+		};
+
+
+		struct Type{
+			Token::ID token;
+		};